fix#2: calculate_time
