+++ app/Console/Commands/CalculateWorkingTime.php
@@ -3,35 +3,35 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
-use App\Interfaces\AbsentInterface;
+use App\Interfaces\TimeLogInterface;
 
-class CaculateAbsentTime extends Command
+class CalculateWorkingTime extends Command
 {
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'caculate:absent-time';
+    protected $signature = 'calculate:working-time';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Caculate Absent Time of Staff';
+    protected $description = 'Caculate Working Time of Staff';
 
-    private $absentRequestRepository;
+    private $timeLogRepository;
     /**
      * Create a new command instance.
      *
      * @return void
      */
-    public function __construct(AbsentInterface $absentRequestRepository)
+    public function __construct(TimeLogInterface $timeLogRepository)
     {
         parent::__construct();
 
-        $this->absentRequestRepository = $absentRequestRepository;
+        $this->timeLogRepository = $timeLogRepository;
     }
 
     /**
@@ -41,6 +41,6 @@ class CaculateAbsentTime extends Command
      */
     public function handle()
     {
-        $this->absentRequestRepository->caculateAbsentTime();
+        $this->timeLogRepository->calculateWorkingTime();
     }
 }
