+++ app/Repositories/AbsentRepository.php
@@ -116,13 +116,13 @@ class AbsentRepository implements AbsentInterface
         ->whereYear('day', $this->operators[$request['operator_year']], $request['year'])->get();
     }
 
-    public function caculateAbsentTime()
+    public function calculateAbsentTime()
     {
         $absents = AbsentRequest::whereNull('absent_time')->get();
 
         foreach ($absents as $absent) {
             $hourDiff = Carbon::parse($absent->time_absent_from)->floatDiffInHours($absent->time_absent_to);
-            
+
             if ($hourDiff >= 8) {
                 $absentTime = 8;
             } else {
@@ -137,6 +137,6 @@ class AbsentRepository implements AbsentInterface
 
     public function getAbsentTime($absents)
     {
-        return array_sum($absents->pluck('absent_time')->toArray());
+        return $absents->sum('absent_time');
     }
 }
